pr_labeler: re-architect triager script (#1882)
This commit reorganizes the issue/PR triager script and updates the
workflow to run more efficiently.

- Make the script a proper Python package instead of an unwieldy single
  file
- Use locked dependencies and UV to decrease workflow runtime to under
  10 seconds.

(cherry picked from commit 7138e42716a7a33fb87d5bf6ce590fa9ec291f68)
(cherry picked from commit 1cf9f7917b6eeac4928fefc2d979ed96cf29664a)
