+++ hacking/pr_labeler/pr_labeler/__main__.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2023 Maxwell G <maxwell@gtmx.me>
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+"""
+Module entrypoint
+"""
+
+from __future__ import annotations
+
+from .cli import APP
+
+if __name__ == "__main__":
+    APP()
