+++ tests/pr_labeler.in
@@ -1,0 +1,1 @@
+../hacking/pr_labeler/requirements.txt
