+++ tests/pr_labeler.txt
@@ -2,10 +2,8 @@
 # This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
-#    pip-compile --allow-unsafe --output-file=tests/typing.txt --strip-extras tests/typing.in
+#    pip-compile --allow-unsafe --output-file=tests/pr_labeler.txt --strip-extras tests/pr_labeler.in
 #
-argcomplete==3.5.0
-    # via nox
 certifi==2024.8.30
     # via requests
 cffi==1.17.1
@@ -17,60 +15,37 @@ charset-normalizer==3.3.2
 click==8.1.7
     # via typer-slim
 codeowners==0.7.0
-    # via -r tests/../hacking/pr_labeler/requirements.txt
-colorlog==6.8.2
-    # via nox
+    # via -r tests/pr_labeler.in
 cryptography==43.0.1
     # via pyjwt
 deprecated==1.2.14
     # via pygithub
-distlib==0.3.8
-    # via virtualenv
-filelock==3.16.0
-    # via virtualenv
-idna==3.9
+idna==3.8
     # via requests
 jinja2==3.1.4
-    # via -r tests/../hacking/pr_labeler/requirements.txt
+    # via -r tests/pr_labeler.in
 markupsafe==2.1.5
     # via jinja2
-mypy==1.11.2
-    # via -r tests/typing.in
-mypy-extensions==1.0.0
-    # via mypy
-nox==2024.4.15
-    # via -r tests/typing.in
-packaging==24.1
-    # via nox
-platformdirs==4.3.3
-    # via virtualenv
 pycparser==2.22
     # via cffi
 pygithub==2.4.0
-    # via -r tests/../hacking/pr_labeler/requirements.txt
+    # via -r tests/pr_labeler.in
 pyjwt==2.9.0
     # via pygithub
 pynacl==1.5.0
     # via pygithub
 requests==2.32.3
     # via pygithub
-tomli==2.0.1
-    # via
-    #   mypy
-    #   nox
 typer-slim==0.12.5
-    # via -r tests/../hacking/pr_labeler/requirements.txt
+    # via -r tests/pr_labeler.in
 typing-extensions==4.12.2
     # via
     #   codeowners
-    #   mypy
     #   pygithub
     #   typer-slim
 urllib3==2.2.3
     # via
     #   pygithub
     #   requests
-virtualenv==20.26.4
-    # via nox
 wrapt==1.16.0
     # via deprecated
